requirements.txt
asgiref==3.5.2
Django==4.1.2
django-versatileimagefield==2.2
Pillow==9.3.0
psycopg2-binary==2.9.5
python-decouple==3.6
python-magic==0.4.27
python-magic-bin==0.4.14
sqlparse==0.4.3
tzdata==2022.6
